Permit to boot more qemus
For the moment only x86_64 qemu are ran.
For booting more qemu, we need to manage the job context.
The easiest way is to manage the context in the test-configs.yaml

So this patch remove the current context part of
generic-qemu-boot-template.jinja2.
All context data will now be in test-configs.yaml

The x86 qemu is kept in place.

The unused arm/arm64 qemu are modified to be more precise in which
machine they emulate.
They are generated with a new device-type naming qemu-arch-machine.
So they will be ignored until labs upgrade a recent LAVA with device
aliases.