kernelci.build: keep .ccache directory inside the build directory
In order to make ccache more likely to be useful, and to be able to
discard it when necessary, keep the .ccache directory inside the
kernel build directory.  That way, it should always cache build data
relevant to the build done in that build directory, and when removing
unused build directories the cache will be removed at the same time.

It's also worth noting that the architecture and compiler are hashed
as part of the ccache data, so if the build directory is not specific
to any arch or compiler the ccache data will not be useful but at
least won't lead to broken builds.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>