baseline: add dmesg tests
Add test cases to the baseline test plan to check for kernel errors
using dmesg.  One test case is created for each log level starting
from "error" and up to "emergency".  Count the number of messages for
each category and report it as a measurement value.  Log messages at
the "warning" level are also checked but not reported as a test case.

All the kernel log lines matching those levels are then output again
in the log for debugging purposes.  The backend can't at the moment
store anything like arbitrary log strings to track individual errors,
so detecting whether there were any errors is a first step.

This relies on the latest version of the "baseline" KernelCI buildroot
file system which has the standard dmesg implementation from
util-linux rather than the busybox one.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>