bisect.jpl: tweak email subject
Remove the git describe to shorten the email subject and move the test
plan name after "bisection:" in preparation for test case bisections.

For example, rather than:

  broonie-sound/for-next boot bisection: v5.4-rc6-292-gc1efaea10be0 on sun8i-h3-libretech-all-h3-cc

we would now have:

  broonie-sound/for-next bisection: boot on sun8i-h3-libretech-all-h3-cc

or potentially:

  broonie-sound/for-next bisection: v4l2-compliance on sun8i-h3-libretech-all-h3-cc

The git describe string was also a bit misleading as it's the "bad"
kernel revision initially tested, not the one found by the bisection.

To better connect this report with the commit it has found, we should
make it reply to the original thread when the patch was submitted on a
mailing list.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>